Debug toNext()
Ensure DOM elements quiz-greeting & question-area are removed at the right time & that question info and results area are populated and at the right time. Info for results area still needs to be added, but w/ the current functionality, it can easily be added at the right time. Next step, however, is to populate questionArea with a function that will be called in the 'if' & 'else' parts of this conditional inside toNext(). No function will be necessary to populate resultsArea, as this will only need to be done under one condition, so this functionality can be put into this conditional statement under the 'else if' condition.